+++ src/components/SquaredGSynAccumulator.cpp
@@ -35,11 +35,11 @@ void SquaredGSynAccumulator::initializeChannelCoefficients() {
 void SquaredGSynAccumulator::updateBufferCPU(double simTime, double deltaTime) {
    float const *gSynExc = mLayerInput->getChannelData(CHANNEL_EXC);
    float *bufferData    = mBufferData.data();
-   int numNeurons       = getBufferSizeAcrossBatch();
+   long numNeurons      = getBufferSizeAcrossBatch();
 #ifdef PV_USE_OPENMP_THREADS
 #pragma omp parallel for schedule(static)
 #endif
-   for (int neuron = 0; neuron < numNeurons; neuron++) {
+   for (long neuron = 0; neuron < numNeurons; neuron++) {
       float gSyn         = gSynExc[neuron];
       bufferData[neuron] = gSyn * gSyn;
    }
